fix:修改获取函数

+++ Device/ANO.h
@@ -10,7 +10,7 @@ void ANO_DT_send_acc(float acc_x, float acc_y, float acc_z);
 
 void ANO_DT_send_vision(int16_t dx, int16_t dy, uint8_t detected);
 
-void ANO_DT_send_encoder(int32_t encoder_count);
+void ANO_DT_send_encoder(uint16_t encoder_count);
 
 void ANO_DT_send_int16(int16_t data1);
